requirements.txt
bcrypt==4.0.1
boto3==1.26.61
botocore==1.29.61
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
circuitbreaker==1.4.0
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.12.0
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.6.3
mosaicml-streaming==0.2.3
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oci==2.90.4
packaging==23.0
paramiko==3.0.0
Pillow==9.4.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tokenizers==0.13.2
torch==1.13.1
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
xxhash==3.2.0
zstd==1.5.2.6
